Revert "Cast quality to int"
This reverts commit ba0d95e2670df15428c7cbbb9f085ca7dea154d5.

+++ src/mako/http/request/Headers.php
@@ -188,7 +188,7 @@ class Headers implements Countable, IteratorAggregate
 
 				// Strip the "q=" part so that we're left with only the numeric value
 
-				$quality = (int) substr(trim($quality), 2);
+				$quality = substr(trim($quality), 2);
 			}
 
 			$groupedAccepts[$quality][] = trim($accept);
